+++ graylog2-server/src/main/java/org/graylog2/bindings/PersistenceServicesBindings.java
@@ -39,7 +39,7 @@ import org.graylog2.security.MongoDBSessionService;
 import org.graylog2.security.MongoDBSessionServiceImpl;
 import org.graylog2.security.ldap.LdapSettingsService;
 import org.graylog2.security.ldap.LdapSettingsServiceImpl;
-import org.graylog2.security.token.AccessTokenServiceImpl;
+import org.graylog2.security.AccessTokenServiceImpl;
 import org.graylog2.shared.users.UserService;
 import org.graylog2.streams.StreamRuleService;
 import org.graylog2.streams.StreamRuleServiceImpl;
