+++ graylog2-server/src/test/java/org/graylog2/security/AccessTokenServiceImplTest.java
@@ -23,9 +23,6 @@ import org.bson.Document;
 import org.bson.types.ObjectId;
 import org.graylog.testing.mongodb.MongoDBFixtures;
 import org.graylog.testing.mongodb.MongoDBInstance;
-import org.graylog2.security.token.AccessTokenCipher;
-import org.graylog2.security.token.AccessTokenImpl;
-import org.graylog2.security.token.AccessTokenServiceImpl;
 import org.joda.time.DateTime;
 import org.joda.time.DateTimeZone;
 import org.junit.After;
